adding initial logging
