+++ src/access/dto/upload.ts
@@ -1,0 +1,10 @@
+import { ApiProperty } from "@nestjs/swagger";
+
+export class UploadDto {
+    @ApiProperty({
+      description: 'Encrypt uploaded data',
+      example: 'false',
+      required: false,
+    })
+    encrypt: string;
+}
