+++ src/shared/nevermined/nvm.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { NeverminedService } from './nvm.service';
 import { ConfigModule } from '../config/config.module';
+import { LoggerModule } from '../logger/logger.module';
 
 @Module({
-  imports: [ConfigModule],
+  imports: [ConfigModule, LoggerModule],
   providers: [NeverminedService],
   exports: [NeverminedService],
 })
